+++ src/ExampleApp.js
@@ -1,34 +1,22 @@
 import React from "react";
 import "./index.css";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import { HomePage } from "./pages/home/HomePage";
-import { News } from "./pages/news/News";
-import { NavBar } from "./components/NavBar";
-import { H1 } from "./components/Text";
+import { BrowserRouter as Router, Route, Switch, Redirect } from "react-router-dom";
+import { HomePage } from "./pages/homeOLD/HomePage";
+import { NewsRouter } from "./pages/news/NewsRouter";
 
-class ExampleApp extends React.Component {
+
+class App extends React.Component {
   render() {
     return (
         <Router>
-          <NavBar/>
           <Switch>
-            <Route path="/" exact>        <News/>                     </Route>
-            <Route path="/arrangement">   <H1>BLOT OG SÅNN!!!</H1>    </Route>
-            <Route path="/bedrift">       <H1>Bedrift</H1>            </Route>
-            <Route path="/internt">       <H1>Internt</H1>            </Route>
-            <Route path="/info">          <H1>Info</H1>               </Route>
-            <Route path="/profil">        <H1>Profil</H1>             </Route>
+            <Route exact path="/"/>
+            <NewsRouter/>
+            <Redirect to="/"/>
           </Switch>
         </Router>
     );
   }
 }
 
-export default ExampleApp;
-
-/* 
-  <>
-    <NavBar />
-    <News />
-  </> 
-      */
+export default App;
